requirements.txt
amqp==2.3.2
billiard==3.5.0.4
celery==4.2.1
certifi==2018.10.15
chardet==3.0.4
Django==2.1.3
django-cors-headers==2.4.0
django-environ==0.4.5
djangorestframework==3.9.0
gunicorn==19.9.0
idna==2.7
kombu==4.2.1
pytz==2018.7
requests==2.20.1
urllib3==1.24.1
vine==1.1.4
